requirements.txt
asgiref                       3.5.2
astroid                       2.12.10
colorama                      0.4.5
dill                          0.3.5.1
Django                        4.1.1
django-cors-headers           3.13.0
djangorestframework           3.13.1
djangorestframework-simplejwt 5.2.0
isort                         5.10.1
lazy-object-proxy             1.7.1
mccabe                        0.7.0
pip                           22.2.1
platformdirs                  2.5.2
psycopg2                      2.9.3
PyJWT                         2.4.0
pylint                        2.15.3
pytz                          2022.2.1
setuptools                    63.2.0
sqlparse                      0.4.2
tomli                         2.0.1
tomlkit                       0.11.4
tzdata                        2022.2
wrapt                         1.14.1

django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1
